requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
google-genai==1.73.1
python-dotenv==1.0.1
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
google-cloud-storage==2.14.0
pydantic-settings==2.5.2
sse-starlette==2.1.0


